Move latest `BgpPeer` and `SwitchPortSettings` from `common` to `nexus-types`
This also addresses some lingering TODOs from #9568 about `From` impls
being defined in the wrong place per RFD 619.
